requirements.txt
flask
gunicorn
google-generativeai
python-dotenv
torch
transformers